// src/app/api/models/ingresso-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface IngressoDto {
  dataIngresso?: string;
  descricaoIngresso?: string;
  idCodigo?: number;
  loCal?: string;
  nomeEvento?: string;
  quantidadeCompra?: number;
  quantidadeIngresso?: number;
  tipoIngresso?: string;
  valorIngresso?: number;
  valorTotal?: number;
}